sidebar in place
In order to get correct height for sticky, add a useEffect(), and a ref to the div containing the sidebar and get the clientBoundingRect.top to get the proper height.  Maybe later figure out how to dynamically add a css class for trhat instead of just inline styles.  Also, be careful because if it is hidden by default on inition render from css in mobile view that useeffect might be incorrect